==> Machine_17/notebooks/00_merge_all_raw_csv.ipynb <==
import os, json, ast
import pandas as pd
import numpy as np
from datetime import datetime
from pathlib import Path

# création d'un dataframe à partir du csv de données
job_events_df = pd.read_csv("../data/raw_unmerged/job_events_17_inf.csv")
job_events_df_sup = pd.read_csv("../data/raw_unmerged/job_events_17-sup.csv")
job_events_df_sup2 = pd.read_csv("../data/raw_unmerged/job_events-17-sup2.csv")
job_events_df.reset_index(level=None, drop=True, inplace=True, col_level=0, col_fill='')
job_events_df.info()

job_events_df_sup.reset_index(level=None, drop=True, inplace=True, col_level=0, col_fill='')
job_events_df_sup.info()

job_events_df_sup2.reset_index(level=None, drop=True, inplace=True, col_level=0, col_fill='')
job_events_df_sup2.info()

job_events_merged_df = pd.concat([job_events_df, job_events_df_sup, job_events_df_sup2], ignore_index=True)
job_events_merged_df = job_events_merged_df.sort_values(by='received_at')

# Vérifier la fusion
job_events_merged_df.head()

job_events_merged_df.tail()

print(len(job_events_merged_df))  

job_events_merged_df.to_csv("../data/raw/job_events.csv", index=False)

metrics_merged_df = pd.read_csv("../data/raw/metrics.csv")
job_events_merged_df = pd.read_csv("../data/raw/job_events.csv")
jobs_merged_df = pd.read_csv("../data/raw/jobs.csv")

# Charger les fichiers CSV
metrics_merged_df = pd.read_csv("../data/raw/metrics.csv")
job_events_merged_df = pd.read_csv("../data/raw/job_events.csv")
jobs_merged_df = pd.read_csv("../data/raw/jobs.csv")

# Convertir les colonnes de date
metrics_merged_df['created_at'] = pd.to_datetime(metrics_merged_df['created_at'])
job_events_merged_df['received_at'] = pd.to_datetime(job_events_merged_df['received_at'])
jobs_merged_df['started_at'] = pd.to_datetime(jobs_merged_df['started_at'])

# Trouver et afficher les valeurs maximales et minimales pour chaque DataFrame
max_date_metrics = metrics_merged_df['created_at'].max()
min_date_metrics = metrics_merged_df['created_at'].min()
print(f"metrics_merged_df - Date Min: {min_date_metrics}, Date Max: {max_date_metrics}")

max_date_job_events = job_events_merged_df['received_at'].max()
min_date_job_events = job_events_merged_df['received_at'].min()
print(f"job_events_merged_df - Date Min: {min_date_job_events}, Date Max: {max_date_job_events}")

max_date_jobs = jobs_merged_df['started_at'].max()
min_date_jobs = jobs_merged_df['started_at'].min()
print(f"jobs_merged_df - Date Min: {min_date_jobs}, Date Max: {max_date_jobs}")


-----------------------------------------------
